remove onRequestTilesOfType onRequestLinkableTiles
These two properties were passed around from the document into
each tile. They were only used by the tile linking hooks.
onRequestTilesOfType was not being used anymore by the linking hooks.
onRequestLinkableTiles is replaced by finding the documentContent
of the TileModel. This approach isn't ideal because of how it ties
these components and hooks to the DocumentContent. I'm working
on a document to describe all the similar types of access throughout
the codebase so we can try to find a consistent approach.
