+++ src/components/document/document-content.tsx
@@ -11,7 +11,6 @@ import { TileRowComponent, kDragResizeRowId, extractDragResizeRowId, extractDrag
 import { DocumentContentModelType } from "../../models/document/document-content";
 import { IDragToolCreateInfo, IDragTilesData } from "../../models/document/document-content-types";
 import { getTileContentInfo } from "../../models/tiles/tile-content-info";
-import { kNoLinkableTiles } from "../../models/tiles/tile-link-types";
 import { getDocumentIdentifier } from "../../models/document/document-utils";
 import { IDropRowInfo } from "../../models/document/tile-row";
 import { logDataTransfer } from "../../models/document/drag-tiles";
@@ -235,8 +234,6 @@ export class DocumentContentComponent extends BaseComponent<IProps, IState> {
                                   documentContent={this.domElement}
                                   rowIndex={index} height={rowHeight} tileMap={tileMap}
                                   dropHighlight={dropHighlight}
-                                  onRequestTilesOfType={this.handleRequestTilesOfType}
-                                  onRequestLinkableTiles={this.handleRequestLinkableTiles}
                                   onRequestUniqueTitle={this.handleRequestUniqueTitle}
                                   ref={(elt) => this.rowRefs.push(elt)} {...others} />
               : null;
@@ -260,28 +257,6 @@ export class DocumentContentComponent extends BaseComponent<IProps, IState> {
     const tile = this.props.content?.getTile(id);
     return tile?.computedTitle;
   }
-
-  // TODO: this should be moved into DocumentContent now that
-  // the document component is not needed to get the title of a tile.
-  // Code currently uses this through onRequestTilesOfType, all of that code
-  // is inside of a tile so it can use the TileModelContext to get the tileModel
-  // then use getDocumentContentFromNode(model) to get the content.
-  private handleRequestTilesOfType = (tileType: string) => {
-    const { content } = this.props;
-    const tileApiInterface = this.context;
-    if (!content || !tileType || !tileApiInterface) return [];
-    const tilesOfType = content.getTilesOfType(tileType);
-    return tilesOfType.map(id => ({ id, title: this.getTileTitle(id) }));
-  };
-
-  // TODO: remove this.
-  // Code currently uses this through onRequestLinkableTiles
-  // Instead it can use the TileModelContext to get the tileModel
-  // then use getDocumentContentFromNode(model) to get the content.
-  private handleRequestLinkableTiles = () => {
-    const { content } = this.props;
-    return content?.getLinkableTiles() || kNoLinkableTiles;
-  };
 
   private handleRequestUniqueTitle = (tileId: string) => {
     const { content } = this.props;
