+++ src/components/document/tile-row.tsx
@@ -4,7 +4,6 @@ import { observer, inject } from "mobx-react";
 import { BaseComponent } from "../base";
 import { TileLayoutModelType, TileRowModelType } from "../../models/document/tile-row";
 import { getTileContentInfo } from "../../models/tiles/tile-content-info";
-import { ILinkableTiles } from "../../models/tiles/tile-link-types";
 import { ITileModel } from "../../models/tiles/tile-model";
 import { SectionHeader } from "../tiles/section-header";
 import { TileApiInterfaceContext } from "../tiles/tile-api";
@@ -68,8 +67,6 @@ interface IProps {
   tileMap: any;
   readOnly?: boolean;
   dropHighlight?: string;
-  onRequestTilesOfType: (tileType: string) => Array<{ id: string, title?: string }>;
-  onRequestLinkableTiles?: () => ILinkableTiles;
   onRequestUniqueTitle: (tileId: string) => string | undefined;
 }
