+++ src/components/tiles/text/spec-text-tile.tsx
@@ -35,9 +35,6 @@ export function specTextTile(options: ISpecTextTileOptions) {
     onSetCanAcceptDrop: (tileId) => {
       throw new Error("Function not implemented.");
     },
-    onRequestTilesOfType: (tileType) => {
-      throw new Error("Function not implemented.");
-    },
     onRequestUniqueTitle: (tileId) => {
       throw new Error("Function not implemented.");
     },
