+++ src/components/tiles/tile-component.test.tsx
@@ -29,8 +29,6 @@ describe("TileComponent", () => {
 
     const onResizeRow = jest.fn();
     const onSetCanAcceptDrop = jest.fn();
-    const onRequestTilesOfType = jest.fn();
-    const onRequestLinkableTiles = jest.fn();
     const onRequestUniqueTitle = jest.fn();
     const onRequestRowHeight = jest.fn();
 
@@ -45,8 +43,6 @@ describe("TileComponent", () => {
             model={toolTileModel}
             onResizeRow={onResizeRow}
             onSetCanAcceptDrop={onSetCanAcceptDrop}
-            onRequestTilesOfType={onRequestTilesOfType}
-            onRequestLinkableTiles={onRequestLinkableTiles}
             onRequestUniqueTitle={onRequestUniqueTitle}
             onRequestRowHeight={onRequestRowHeight}
             />
