+++ src/components/tiles/tile-component.tsx
@@ -6,7 +6,6 @@ import { isAlive } from "mobx-state-tree";
 import ResizeObserver from "resize-observer-polyfill";
 import { transformCurriculumImageUrl } from "../../models/tiles/image/image-import-export";
 import { getTileComponentInfo } from "../../models/tiles/tile-component-info";
-import { ILinkableTiles } from "../../models/tiles/tile-link-types";
 import { ITileModel } from "../../models/tiles/tile-model";
 import { BaseComponent } from "../base";
 import PlaceholderTileComponent from "./placeholder/placeholder-tile";
@@ -68,8 +67,6 @@ interface ITileBaseProps {
   readOnly?: boolean;
   onResizeRow: (e: React.DragEvent<HTMLDivElement>) => void;
   onSetCanAcceptDrop: (tileId?: string) => void;
-  onRequestTilesOfType: (tileType: string) => Array<{ id: string, title?: string }>;
-  onRequestLinkableTiles?: () => ILinkableTiles;
   // TODO: this isn't really necessary to be in component API anymore, it is
   // implemented in the model layer.
   onRequestUniqueTitle: (tileId: string) => string | undefined;
