+++ src/hooks/use-provider-tile-linking.tsx
@@ -1,25 +1,23 @@
 import { useCallback } from "react";
 
-import {
-  ILinkableTiles, ITileLinkMetadata, ITypedTileLinkMetadata, kNoLinkableTiles
+import { ITileLinkMetadata
 } from "../models/tiles/tile-link-types";
 import { ITileModel } from "../models/tiles/tile-model";
 import { useLinkProviderTileDialog } from "./use-link-provider-tile-dialog";
 import { getTileContentById } from "../utilities/mst-utils";
 import { SharedDataSet } from "../models/shared/shared-data-set";
+import { useLinkableTiles } from "./use-linkable-tiles";
 
 interface IProps {
   actionHandlers?: any;
   model: ITileModel;
   readOnly?: boolean;
-  onRequestTilesOfType: (tileType: string) => ITileLinkMetadata[];
-  onRequestLinkableTiles?: () => ILinkableTiles;
 }
 export const useProviderTileLinking = ({
-  actionHandlers, model, readOnly, onRequestTilesOfType, onRequestLinkableTiles
+  actionHandlers, model, readOnly
 }: IProps) => {
   const {handleRequestTileLink, handleRequestTileUnlink} = actionHandlers || {};
-  const { providers: linkableTiles } = useLinkableTiles({ model, onRequestTilesOfType, onRequestLinkableTiles });
+  const { providers: linkableTiles } = useLinkableTiles({ model });
   const isLinkEnabled = (linkableTiles.length > 0);
 
   const linkTile = useCallback((tileInfo: ITileLinkMetadata) => {
@@ -55,29 +53,4 @@ export const useProviderTileLinking = ({
           });
 
   return { isLinkEnabled, showLinkTileDialog };
-};
-
-interface IUseLinkableTilesProps {
-  model: ITileModel;
-  onRequestTilesOfType: (tileType: string) => ITileLinkMetadata[];
-  onRequestLinkableTiles?: () => ILinkableTiles;
-}
-const useLinkableTiles = ({ model, onRequestTilesOfType, onRequestLinkableTiles }: IUseLinkableTilesProps) => {
-  const { providers, consumers } = onRequestLinkableTiles?.() || kNoLinkableTiles;
-
-  // add default title if there isn't a title
-  const countsOfType = {} as Record<string, number>;
-  function addDefaultTitle({ id, type, title, titleBase }: ITypedTileLinkMetadata) {
-    if (!countsOfType[type]) {
-      countsOfType[type] = 1;
-    } else {
-      countsOfType[type]++;
-    }
-    return { id, type, title: title || `${titleBase || type} ${countsOfType[type]}`};
-  }
-
-  return {
-    providers: providers.map(addDefaultTitle),
-    consumers: consumers.map(addDefaultTitle)
-  };
 };
