+++ src/plugins/data-card/data-card-tile.test.tsx
@@ -39,9 +39,6 @@ describe("DataCardToolComponent", () => {
     onSetCanAcceptDrop: (tileId?: string): void => {
       throw new Error("Function not implemented.");
     },
-    onRequestTilesOfType: (tileType: string): { id: string; title?: string | undefined; }[] => {
-      throw new Error("Function not implemented.");
-    },
     onRequestUniqueTitle: (tileId: string): string | undefined => {
       return "Data Card Collection 1";
     },
